requirements.txt
flask==2.3.3
requests==2.31.0
pyyaml==6.0.1
simple-justwatch-python-api==1.1.0
schedule==1.2.0
python-dateutil==2.8.2
